Simplified Instance creation
